refactor(gts-id): modularize id types and tighten the public API
Split the monolithic gts-id crate into focused modules (parse, gts_id,
gts_id_segment, gts_wildcard, error). GtsID / GtsWildcard fields are now
private behind accessors (id(), segments(), into_segments()), and the
error type is the GtsIdError struct.

Move deterministic UUID v5 derivation into gts-id as a GtsID::to_uuid
inherent method, gated behind the crate's optional `uuid` feature (off by
default). The gts crate enables the feature; gts-macros depends on gts-id
without it, so uuid stays out of the proc-macro build graph and out of
gts-id's default public API. This replaces the former GtsIdUuidExt
extension trait that previously lived in the gts crate.

Adapt the dependent crates to the tightened API: field access -> accessor
methods, gts_id_segments -> segments(), GtsError -> GtsIdError. gts.rs is
slimmed to re-export the id primitives from gts-id and keep the
schema-aware GtsTypeId / GtsInstanceId wrappers.
