+++ gts-macros/tests/compile_fail/instance_id_invalid_format.stderr
@@ -1,4 +1,4 @@
-error: Invalid GTS instance ID: segment #2: Too few tokens (got 4, min 5). Expected format: vendor.package.namespace.type.vMAJOR[.MINOR]
+error: Invalid GTS instance ID: Invalid GTS segment #2 @ offset 27: 'not.a.valid.segment': Too few tokens (got 4, min 5). Expected format: vendor.package.namespace.type.vMAJOR[.MINOR]
   --> tests/compile_fail/instance_id_invalid_format.rs:26:13
    |
 26 |         id: "gts.acme.core.test.perm.v1~not.a.valid.segment",
