+++ gts-macros/tests/compile_fail/instance_raw_no_tilde.stderr
@@ -1,4 +1,4 @@
-error: instance id literal must contain at least one `~` (chained form: gts.<type>~<instance>)
+error: Invalid GTS instance ID: Invalid GTS identifier: gts.acme.core.events.topic.v1: Single-segment instance IDs are prohibited. Instance IDs must be chained with at least one type segment (e.g., 'type~instance')
   --> tests/compile_fail/instance_raw_no_tilde.rs:11:15
    |
 11 |         "id": "gts.acme.core.events.topic.v1",
