+++ gts-macros/tests/compile_fail/invalid_gts_id_too_many_tokens.stderr
@@ -1,4 +1,4 @@
-error: struct_to_gts_schema: Segment #2: Too many name tokens before version (got 5, expected 4). Expected format: vendor.package.namespace.type.vMAJOR[.MINOR]
+error: struct_to_gts_schema: Invalid GTS type ID: Invalid GTS segment #2 @ offset 28: 'x.core.license_enforcer.integration.plugin.v1~': Too many name tokens before version (got 5, expected 4). Expected format: vendor.package.namespace.type.vMAJOR[.MINOR]
   --> tests/compile_fail/invalid_gts_id_too_many_tokens.rs:25:15
    |
 25 |     type_id = "gts.x.core.modkit.plugin.v1~x.core.license_enforcer.integration.plugin.v1~",
