+++ gts/Cargo.toml
@@ -16,11 +16,10 @@ publish = true
 workspace = true
 
 [dependencies]
-gts-id.workspace = true
+gts-id = { workspace = true, features = ["uuid"] }
 serde.workspace = true
 serde_json.workspace = true
 thiserror.workspace = true
-uuid.workspace = true
 jsonschema.workspace = true
 schemars.workspace = true
 walkdir.workspace = true
