+++ gts/src/lib.rs
@@ -19,7 +19,7 @@ pub use entities::{GtsConfig, GtsEntity, GtsFile, ValidationError, ValidationRes
 pub use files_reader::GtsFileReader;
 #[allow(deprecated)]
 pub use gts::GtsSchemaId;
-pub use gts::{GtsError, GtsID, GtsIdSegment, GtsInstanceId, GtsTypeId, GtsWildcard};
+pub use gts::{GtsID, GtsIdError, GtsIdSegment, GtsInstanceId, GtsTypeId, GtsWildcard};
 pub use ops::GtsOps;
 pub use path_resolver::JsonPathResolver;
 pub use schema::{
